Initialize App data from firestore
Redux is now initialized from firestore. We need to do a bit more work on the DraggableWorkoutList since it was displaying the exerciseIds and not the names.

Also I left color out of Firstore for now since we haven't decided what we want to do for that. But I'll add that in a later PR if that's what we decide to go with.